added players context
