+++ src/contexts/players-context.tsx
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import type { PlayersContextType } from "@/lib/types.ts";
+
+export const PlayersContext = createContext<PlayersContextType | undefined>(
+  undefined,
+);
